Fixed my lazy typescript conversions Did some light refactoring in the process. Need to do more...

+++ src/replace-servers.ts
@@ -4,55 +4,60 @@ import { color } from "./lib/lunLib.js";
 /** @param {NS} ns */
 export async function main(ns:NS) {
 	var i = 0;
-	let ram:any;
-	ram = ns.args[0];
+	let ram = ns.args[0];
 	var skipprompt = ns.args[1];
 	var serverlimit = ns.getPurchasedServerLimit();
-	var fcost = ns.nFormat((ns.getPurchasedServerCost(ram) * serverlimit), "$0.000a");
-
+	
 	if (!ram) {
 		ns.tprint(`${color.red}Error: No RAM value specified.`);
 		ns.exit();
 	}
 
-	if (!Number.isInteger(Math.log2(ram))) {
-		ns.tprint(`${color.red}Error: Invalid RAM value specified.`);
-		ns.exit();
-	}
+	if (typeof ram === "number") {
+		var fcost = ns.nFormat((ns.getPurchasedServerCost(ram) * serverlimit), "$0.000a");
 
-	if (ram > ns.getPurchasedServerMaxRam()) {
-		ns.tprint(`${color.orange}Alert: Specified RAM is too large to be purchased.`)
-		ns.exit();
-	}
-	if (!ns.serverExists("pserv-0")) {
+		if (!Number.isInteger(Math.log2(ram))) {
+			ns.tprint(`${color.red}Error: Invalid RAM value specified.`);
+			ns.exit();
+		}
+
+		if (ram > ns.getPurchasedServerMaxRam()) {
+			ns.tprint(`${color.orange}Alert: Specified RAM is too large to be purchased.`)
+			ns.exit();
+		}
+		if (!ns.serverExists("pserv-0")) {
+			if (!skipprompt) {
+				if (!await ns.prompt(`Purchasing ${serverlimit} servers with ${ram}GB will cost ${fcost}. Are you sure?`)) {
+					ns.tprint("Purchase cancelled.");
+					ns.exit();
+				}
+			}
+			ns.run("purchase-servers.js", 1, ram, true);
+			ns.exit();
+		}
+
+
+		if (ram <= ns.getServerMaxRam("pserv-0")) {
+			ns.tprint(`${color.white}Info: The RAM specified is not higher than what is currently being used.`);
+			ns.tprint(`${color.white}Server array is currently running on ${ns.getServerMaxRam("pserv-0")}GB per server.`);
+			ns.exit();
+		}
 		if (!skipprompt) {
-			if (!await ns.prompt(`Purchasing ${serverlimit} servers with ${ram}GB will cost ${fcost}. Are you sure?`)) {
+			if (!await ns.prompt(`Replacing ${serverlimit} servers for ${ram}GB will cost ${fcost}. Are you sure?`)) {
 				ns.tprint("Purchase cancelled.");
 				ns.exit();
 			}
 		}
-		ns.run("purchase-servers.js", 1, ram, true);
-		ns.exit();
-	}
 
-
-	if (ram <= ns.getServerMaxRam("pserv-0")) {
-		ns.tprint(`${color.white}Info: The RAM specified is not higher than what is currently being used.`)
-		ns.tprint(`${color.white}Server array is currently running on ${ns.getServerMaxRam("pserv-0")}GB per server.`)
-		ns.exit();
-	}
-	if (!skipprompt) {
-		if (!await ns.prompt(`Replacing ${serverlimit} servers for ${ram}GB will cost ${fcost}. Are you sure?`)) {
-			ns.tprint("Purchase cancelled.");
-			ns.exit();
+		for (i = 0; i < serverlimit; i++) {
+			ns.killall("pserv-" + i);
+			ns.deleteServer("pserv-" + i);
+			await ns.sleep(50);
 		}
-	}
 
-	for (i = 0; i < serverlimit; i++) {
-		ns.killall("pserv-" + i);
-		ns.deleteServer("pserv-" + i);
-		await ns.sleep(50);
+		ns.run("purchase-servers.js", 1, ram, true);
+	} else {
+		ns.tprint(`${color.red}Error: RAM must be entered as a number.`);
+		ns.exit()
 	}
-
-	ns.run("purchase-servers.js", 1, ram, true);
 }
